@@ -19,7 +19,7 @@
     "config": {
         "sort-packages": true,
         "platform": {
-            "php": "7.4.99"
+            "php": "8.0.99"
         },
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true
@@ -32,7 +32,7 @@
         }
     },
     "require": {
-        "php": ">=7.4, <8.2",
+        "php": "~8.0.0 || ~8.1.0 || ~8.2.0",
         "laminas/laminas-stdlib": "^3.6",
         "psr/container": "^1.1.1",
         "psr/log": "^1.1.4 || ^3.0.0"
@@ -49,8 +49,7 @@
     "conflict": {
         "zendframework/zend-di": "*",
         "laminas/laminas-servicemanager-di": "*",
-        "laminas/laminas-servicemanager": "<3.13.0",
-        "phpspec/prophecy": "<1.9.0"
+        "laminas/laminas-servicemanager": "<3.13.0"
     },
     "suggest": {
         "laminas/laminas-servicemanager": "An IoC container without auto wiring capabilities"
