@@ -6,8 +6,6 @@ namespace Laminas\Di;
 
 use Psr\Container\ContainerInterface;
 
-use function get_class;
-
 /**
  * Default IoC container implementation.
  *
@@ -35,7 +33,7 @@ class DefaultContainer implements ContainerInterface
 
         $this->services[InjectorInterface::class]  = $injector;
         $this->services[ContainerInterface::class] = $this;
-        $this->services[get_class($injector)]      = $injector;
+        $this->services[$injector::class]          = $injector;
         $this->services[static::class]             = $this;
     }
 
